+++ src/test/java/org/folio/rest/support/http/InterfaceUrls.java
@@ -134,6 +134,10 @@ public class InterfaceUrls {
     return StorageTestSuite.storageUrl("/nature-of-content-terms" + subPath);
   }
 
+  public static URL authoritySourceFilesUrl(String subPath) {
+    return StorageTestSuite.storageUrl("/authority-source-files" + subPath);
+  }
+
   public static URL authorityNoteTypesUrl(String subPath) {
     return StorageTestSuite.storageUrl("/authority-note-types" + subPath);
   }