+++ NEWS.md
@@ -1,5 +1,6 @@
 ## 25.0.0 IN-PROGRESS
 
+* Create a pre-defined Authority Source file list (MODINVSTOR-892)
 * Fixed effective location migration script for holdings records (MODINVSTOR-940)
 * Adds integrity checks for statistical code types during upgrade ([MODINVSTOR-935] (https://issues.folio.org/browse/MODINVSTOR-935))
 * GET Instance Set by CQL ([MODINVSTOR-918](https://issues.folio.org/browse/MODINVSTOR-918))